Update Things Page : Renamed DesktopSmallBox Filename
